split existing JS into modules
