+++ src/js/main.js
@@ -1,2 +1,92 @@
-require('./org.js');
-require('./vol.js');
+import {signOut} from './org/auth';
+import {toggleFlexible} from './org/opp-forms';
+import {idInArray} from './vol/helpers';
+import {loadStore, saveOpportunity, removeOppFromStore} from './vol/local-storage';
+import {selectAllCategories, toggleSelectAllCheckbox} from './vol/filters-form';
+import {updateSaveButton, renderSavedOpps} from './vol/dom-saved-opps';
+
+// don't run this outside of a browser environment (e.g., when testing in Node)
+if (typeof window !== "undefined") {
+
+  // Code in this anonymous function is immediately invoked once this script loads:
+  (()=> {
+  
+    // get data from localStorage:  
+    let store = loadStore();
+
+    // create variables from DOM elements
+    // if a given element isn't found on the current page, its variable has value of null
+    const signOutLink = document.querySelectorAll(".signOutLink");
+    const flexibleCheckbox = document.getElementById("flexible");
+    const selectAllButton = document.getElementById("selectAll");
+    const categoryInputs = document.querySelectorAll("input[name=category]");
+    const saveButton = document.getElementById("save-button");
+    const oppListParent = document.getElementById("opp-container");
+
+    // when the sign-out link is clicked, the signOut function is executed
+    if (signOutLink) {
+      signOutLink.forEach(link => {
+        link.addEventListener("click", signOut);
+      });
+    }
+
+    // set up Add/Edit opp forms:
+    if (flexibleCheckbox) {
+      flexibleCheckbox.addEventListener("click", toggleFlexible);
+    }
+
+    // set up Filters page:
+    if (selectAllButton) {
+      selectAllButton.addEventListener("change", selectAllCategories);
+      categoryInputs.forEach(cat => {
+        cat.addEventListener("change", toggleSelectAllCheckbox);
+      });
+    }
+
+    // set up Browse Opportunities page:
+    if (saveButton) {
+      // update the save button if the opp is already saved on page load
+      // (currentOpp is a global variable loaded in the HTML created by the Jinja template)
+
+      if (idInArray(store, currentOpp)) {
+        updateSaveButton.bind(saveButton)(true);
+      }
+
+      // on Save Button click, add the opp to localStorage and change button appearance
+      saveButton.addEventListener("click", () => {
+        // depending on if the opportunity has already been saved, either save or remove it:
+        if (!idInArray(store, currentOpp)) {
+          saveOpportunity(store, currentOpp);
+          updateSaveButton.bind(this)(true);
+          store = loadStore();
+        } else {
+          removeOppFromStore(currentOpp.id);
+          updateSaveButton.bind(this)(false);
+          store = loadStore();
+        }
+      });
+    }
+
+    // set up My Saved Opportunities page:
+    if (oppListParent) {
+      // initial render
+      renderSavedOpps(store, oppListParent);
+
+      // listen for clicks on Remove buttons
+      oppListParent.addEventListener("click", e => {
+        if (
+          e.target.classList.contains("remove-button") ||
+          e.target.parentNode.classList.contains("remove-button")
+        ) {
+          const clickedOppId =
+            e.target.parentNode.parentNode.parentNode.parentNode.dataset.id;
+          removeOppFromStore(clickedOppId);
+
+          // re-load the store from localStorage, and re-render the view
+          store = loadStore();
+          renderSavedOpps(store, oppListParent);
+        }
+      });
+    }
+  })();
+}
